Delete timeline.js file
All of this js file was actually commented out in this version because it was a trial and error scratch sheet. Because of that it doesn't affect the functionality of the site at all, and didn't end up getting used because I can pull the JSON directly via the script tags in HTML. JS file not needed in this case because the rest of the timeline is linked to the Knightlab open source!